Replace the tokenizer with a decoder preference
