+++ super-csv/src/main/java/org/supercsv/io/CsvBeanReader.java
@@ -24,6 +24,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.supercsv.cellprocessor.ift.CellProcessor;
+import org.supercsv.decoder.CsvDecoder;
 import org.supercsv.exception.SuperCsvConstraintViolationException;
 import org.supercsv.exception.SuperCsvException;
 import org.supercsv.exception.SuperCsvReflectionException;
@@ -62,21 +63,6 @@ public class CsvBeanReader extends AbstractCsvReader implements ICsvBeanReader {
 	 */
 	public CsvBeanReader(final Reader reader, final CsvPreference preferences) {
 		super(reader, preferences);
-	}
-	
-	/**
-	 * Constructs a new <tt>CsvBeanReader</tt> with the supplied (custom) Tokenizer and CSV preferences. The tokenizer
-	 * should be set up with the Reader (CSV input) and CsvPreference beforehand.
-	 * 
-	 * @param tokenizer
-	 *            the tokenizer
-	 * @param preferences
-	 *            the CSV preferences
-	 * @throws NullPointerException
-	 *             if tokenizer or preferences are null
-	 */
-	public CsvBeanReader(final ITokenizer tokenizer, final CsvPreference preferences) {
-		super(tokenizer, preferences);
 	}
 	
 	/**
