+++ super-csv/src/main/java/org/supercsv/util/EmptyColumnParsing.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.supercsv.io;
+package org.supercsv.util;
 
 /**
  * @author Pietro Aragona
