+++ super-csv/src/test/java/org/supercsv/io/AbstractCsvReaderTest.java
@@ -28,8 +28,11 @@ import java.util.Arrays;
 import java.util.List;
 
 import org.junit.After;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.supercsv.comment.CommentMatches;
+import org.supercsv.comment.CommentStartsWith;
 import org.supercsv.exception.SuperCsvException;
 import org.supercsv.prefs.CsvPreference;
 
@@ -52,20 +55,12 @@ public class AbstractCsvReaderTest {
 	
 	private AbstractCsvReader abstractReader;
 	
-	private AbstractCsvReader tokenizerAbstractReader;
-	
 	private AbstractCsvReader surroundingSpacesNeedQuotesAbstractReader;
-	
-	private ITokenizer tokenizer;
 	
 	/**
 	 * Implementation of AbstractCsvReader for testing.
 	 */
 	private static class MockCsvReader extends AbstractCsvReader {
-		
-		public MockCsvReader(ITokenizer tokenizer, CsvPreference preferences) {
-			super(tokenizer, preferences);
-		}
 		
 		public MockCsvReader(Reader reader, CsvPreference preferences) {
 			super(reader, preferences);
@@ -82,9 +77,6 @@ public class AbstractCsvReaderTest {
 			+ "Harry,Potter,,\"Gryffindor\nHogwarts Castle\nUK\"");
 		abstractReader = new MockCsvReader(reader, PREFS);
 		
-		tokenizer = new Tokenizer(reader, PREFS);
-		tokenizerAbstractReader = new MockCsvReader(tokenizer, PREFS);
-		
 		surroundingSpacesNeedQuotesReader = new StringReader("firstName, lastName, age, address\n"
 			+ " John , Smith, 23 , \n" + "Harry, Potter, , \"Gryffindor\nHogwarts Castle\nUK\" ");
 		surroundingSpacesNeedQuotesAbstractReader = new MockCsvReader(surroundingSpacesNeedQuotesReader,
@@ -97,7 +89,6 @@ public class AbstractCsvReaderTest {
 	@After
 	public void tearDown() throws IOException {
 		abstractReader.close();
-		tokenizerAbstractReader.close();
 		surroundingSpacesNeedQuotesAbstractReader.close();
 	}
 	
@@ -110,15 +101,6 @@ public class AbstractCsvReaderTest {
 	}
 	
 	/**
-	 * Tests a normal reading scenario (with the custom tokenizer reader), asserting all of the properties available
-	 * each time.
-	 */
-	@Test
-	public void testReadingWithTokenizerReader() throws IOException {
-		assertReading(tokenizerAbstractReader);
-	}
-	
-	/**
 	 * Reusable method to test a normal reading scenario, asserting all of the properties are available each time.
 	 */
 	private void assertReading(final AbstractCsvReader csvReader) throws IOException {
@@ -127,7 +109,7 @@ public class AbstractCsvReaderTest {
 		
 		assertEquals(0, csvReader.getLineNumber());
 		assertEquals(0, csvReader.getRowNumber());
-		assertEquals("", csvReader.getUntokenizedRow());
+		assertEquals("", csvReader.getUndecodedRow());
 		assertEquals(0, csvReader.length());
 		
 		// read the header
@@ -140,7 +122,7 @@ public class AbstractCsvReaderTest {
 		
 		assertEquals(1, csvReader.getLineNumber());
 		assertEquals(1, csvReader.getRowNumber());
-		assertEquals("firstName,lastName,age,address", csvReader.getUntokenizedRow());
+		assertEquals("firstName,lastName,age,address", csvReader.getUndecodedRow());
 		assertEquals(4, csvReader.length());
 		
 		// read the first data row
@@ -158,7 +140,7 @@ public class AbstractCsvReaderTest {
 		
 		assertEquals(2, csvReader.getLineNumber());
 		assertEquals(2, csvReader.getRowNumber());
-		assertEquals("John,Smith,23,", csvReader.getUntokenizedRow());
+		assertEquals("John,Smith,23,", csvReader.getUndecodedRow());
 		assertEquals(4, csvReader.length());
 		
 		// read the second data row
@@ -176,14 +158,14 @@ public class AbstractCsvReaderTest {
 		
 		assertEquals(5, csvReader.getLineNumber()); // 2 newlines in harry's address
 		assertEquals(3, csvReader.getRowNumber());
-		assertEquals("Harry,Potter,,\"Gryffindor\nHogwarts Castle\nUK\"", csvReader.getUntokenizedRow());
+		assertEquals("Harry,Potter,,\"Gryffindor\nHogwarts Castle\nUK\"", csvReader.getUndecodedRow());
 		assertEquals(4, csvReader.length());
 		
 		// read again (should be EOF)
 		assertFalse(csvReader.readRow());
 		assertEquals(5, csvReader.getLineNumber());
 		assertEquals(3, csvReader.getRowNumber());
-		assertEquals("", csvReader.getUntokenizedRow());
+		assertEquals("", csvReader.getUndecodedRow());
 		assertEquals(0, csvReader.length());
 		
 	}
@@ -207,7 +189,7 @@ public class AbstractCsvReaderTest {
 		
 		assertEquals(0, csvReader.getLineNumber());
 		assertEquals(0, csvReader.getRowNumber());
-		assertEquals("", csvReader.getUntokenizedRow());
+		assertEquals("", csvReader.getUndecodedRow());
 		assertEquals(0, csvReader.length());
 		
 		// read the header
@@ -220,7 +202,7 @@ public class AbstractCsvReaderTest {
 		
 		assertEquals(1, csvReader.getLineNumber());
 		assertEquals(1, csvReader.getRowNumber());
-		assertEquals("firstName, lastName, age, address", csvReader.getUntokenizedRow());
+		assertEquals("firstName, lastName, age, address", csvReader.getUndecodedRow());
 		assertEquals(4, csvReader.length());
 		
 		// read the first data row
@@ -238,7 +220,7 @@ public class AbstractCsvReaderTest {
 		
 		assertEquals(2, csvReader.getLineNumber());
 		assertEquals(2, csvReader.getRowNumber());
-		assertEquals(" John , Smith, 23 , ", csvReader.getUntokenizedRow());
+		assertEquals(" John , Smith, 23 , ", csvReader.getUndecodedRow());
 		assertEquals(4, csvReader.length());
 		
 		// read the second data row
@@ -256,14 +238,14 @@ public class AbstractCsvReaderTest {
 		
 		assertEquals(5, csvReader.getLineNumber()); // 2 newlines in harry's address
 		assertEquals(3, csvReader.getRowNumber());
-		assertEquals("Harry, Potter, , \"Gryffindor\nHogwarts Castle\nUK\" ", csvReader.getUntokenizedRow());
+		assertEquals("Harry, Potter, , \"Gryffindor\nHogwarts Castle\nUK\" ", csvReader.getUndecodedRow());
 		assertEquals(4, csvReader.length());
 		
 		// read again (should be EOF)
 		assertFalse(csvReader.readRow());
 		assertEquals(5, csvReader.getLineNumber());
 		assertEquals(3, csvReader.getRowNumber());
-		assertEquals("", csvReader.getUntokenizedRow());
+		assertEquals("", csvReader.getUndecodedRow());
 		assertEquals(0, csvReader.length());
 		
 	}
@@ -315,23 +297,309 @@ public class AbstractCsvReaderTest {
 	public void testReaderConstructorWithNullPreferences() {
 		new CsvListReader(reader, null);
 	}
-	
+
 	/**
-	 * Tests the Tokenizer constructor with a null Reader.
+	 * Tests the readRow() method when a newline is reached in quote
+	 * scoped when a single line is only supposed to be read
 	 */
-	@SuppressWarnings("resource")
-	@Test(expected = NullPointerException.class)
-	public void testTokenizerConstructorWithNullReader() {
-		new CsvListReader((Tokenizer) null, PREFS);
+	@Test
+	public void testQuotedFieldWithUnexpectedNewline() throws Exception {
+
+		// Row 2 has a missing trailing quote
+		final String input = "col1,col2\n" +
+				"\"foo\",\"bar\n" +
+				"\"baz\",\"zoo\"\n" +
+				"\"aaa\",\"bbb\"";
+		CsvPreference pref = new CsvPreference.Builder(CsvPreference.EXCEL_PREFERENCE)
+				.maxLinesPerRow(1).build();
+
+		AbstractCsvReader reader = new MockCsvReader(new StringReader(input), pref);
+		try {
+			boolean first = reader.readRow();
+			assertEquals(true , first);
+
+			reader.readRow();
+			fail("should have thrown SuperCsvException");
+		}
+		catch(SuperCsvException e) {
+			assertEquals("unexpected end of line while reading quoted column on line 2",
+					e.getMessage());
+		}
+		reader.close();
 	}
-	
+
+	@Test
+	public void testQuotedFieldWithUnexpectedNewlineNoNextLineRead() throws Exception {
+
+		// Row 2 has a missing trailing quote
+		final String input = "col1,col2\n" +
+				"\"foo\",\"bar\n" +
+				"\"baz\",\"zoo\"\n" +
+				"\"aaa\",\"bbb\"";
+		CsvPreference pref = new CsvPreference.Builder(CsvPreference.EXCEL_PREFERENCE)
+				.maxLinesPerRow(1).build();
+
+		AbstractCsvReader reader = new MockCsvReader(new StringReader(input), pref);
+
+		try {
+			final boolean first = reader.readRow();
+			assertEquals(true , first);
+			assertEquals("[col1, col2]" , reader.getColumns().toString());
+
+			reader.readRow();
+			fail("should have thrown SuperCsvException");
+		}
+		catch(SuperCsvException e) {
+			assertEquals("unexpected end of line while reading quoted column on line 2",
+					e.getMessage());
+		}
+		final boolean third = reader.readRow();
+		assertEquals(true , third);
+		assertEquals("[baz, zoo]" , reader.getColumns().toString());
+
+		final boolean fourth = reader.readRow();
+		assertEquals(true , fourth);
+		assertEquals("[aaa, bbb]" , reader.getColumns().toString());
+
+		//line 4 was the last
+		final boolean fifth = reader.readRow();
+		assertEquals(false , fifth);
+
+		reader.close();
+	}
+
 	/**
-	 * Tests the Tokenizer constructor with a null preference.
+	 * Tests the readRow() method when a newline is reached in quote
+	 * scoped when two lines are only supposed to be read
 	 */
-	@SuppressWarnings("resource")
-	@Test(expected = NullPointerException.class)
-	public void testTokenizerConstructorWithNullPreferences() {
-		new CsvListReader(tokenizer, null);
+	@Test
+	public void testQuotedFieldWithTwoMaxLines() throws Exception {
+
+		// Row 2 has a missing trailing quote
+		final String input = "col1,col2\n" +
+				"\"foo\",\"bar\n" +
+				"baz,zoo\n" +
+				"aaa,bbb";
+		CsvPreference pref = new CsvPreference.Builder(CsvPreference.EXCEL_PREFERENCE)
+				.maxLinesPerRow(2).build();
+
+		AbstractCsvReader reader = new MockCsvReader(new StringReader(input), pref);
+		try {
+			boolean first = reader.readRow();
+			assertEquals(true , first);
+
+			boolean second = reader.readRow();
+			assertEquals(true , second);
+
+			reader.readRow();
+			fail("should have thrown SuperCsvException");
+		}
+		catch(SuperCsvException e) {
+			assertEquals("max number of lines to read exceeded while reading quoted column beginning on line 2 and ending on line 3",
+					e.getMessage());
+		}
+		reader.close();
 	}
-	
+
+	/**
+	 * Tests the readRow() method when a newline is reached in quote
+	 * scoped when two lines are only supposed to be read
+	 */
+	@Test
+	public void testQuotedFieldWithTwoMaxLinesNoMoreLinesRead() throws Exception {
+
+		// Row 2 has a missing trailing quote
+		final String input = "col1,col2\n" +
+				"\"foo,bar\n" +
+				"baz,zoo\n" +
+				"aaa,bbb";
+		CsvPreference pref = new CsvPreference.Builder(CsvPreference.EXCEL_PREFERENCE)
+				.maxLinesPerRow(2).build();
+
+		AbstractCsvReader reader = new MockCsvReader(new StringReader(input), pref);
+		try {
+			boolean first = reader.readRow();
+			assertEquals(true , first);
+			assertEquals("[col1, col2]" , reader.getColumns().toString());
+
+
+			boolean second = reader.readRow();
+			assertEquals(true , second);
+			assertEquals("[\"foo,bar]" , reader.getColumns().toString());
+
+
+			reader.readRow();
+			fail("should have thrown SuperCsvException");
+		}
+		catch(SuperCsvException e) {
+			assertEquals("max number of lines to read exceeded while reading quoted column beginning on line 2 and ending on line 3",
+					e.getMessage());
+		}
+		boolean fourth = reader.readRow();
+		assertEquals(true , fourth);
+		assertEquals("[aaa, bbb]" , reader.getColumns().toString());
+
+		reader.close();
+	}
+
+	/**
+	 * Tests that the CommentMatches comment matcher works (comments are skipped).
+	 */
+	@Test
+	public void testSkipCommentsMatches() throws IOException {
+
+		final CsvPreference commentsMatchesPrefs = new CsvPreference.Builder(CsvPreference.EXCEL_PREFERENCE
+		).skipComments(new CommentMatches("<!--.*-->")).build();
+
+		final String input = "<!--comment-->\nnot,a,comment\n<!-- another comment-->\nalso,not,comment";
+		AbstractCsvReader reader = new MockCsvReader(new StringReader(input), commentsMatchesPrefs);
+
+		reader.readRow();
+		assertTrue(reader.getColumns().size() == 3);
+		assertEquals("not", reader.getColumns().get(0));
+		assertEquals("a", reader.getColumns().get(1));
+		assertEquals("comment", reader.getColumns().get(2));
+
+		reader.readRow();
+		assertTrue(reader.getColumns().size() == 3);
+		assertEquals("also", reader.getColumns().get(0));
+		assertEquals("not", reader.getColumns().get(1));
+		assertEquals("comment", reader.getColumns().get(2));
+
+		assertFalse(reader.readRow());
+
+		reader.close();
+	}
+
+
+	/**
+	 * Tests that the CommentStartsWith comment matcher works (comments are skipped).
+	 */
+	@Test
+	public void testSkipCommentsStartsWith() throws IOException {
+
+		final CsvPreference commentsStartWithPrefs = new CsvPreference.Builder(CsvPreference.EXCEL_PREFERENCE).skipComments(
+				new CommentStartsWith("#")).build();
+
+		final String input = "#comment\nnot,a,comment\n# another comment\nalso,not,comment";
+		AbstractCsvReader reader = new MockCsvReader(new StringReader(input), commentsStartWithPrefs);
+		reader.readRow();
+		assertTrue(reader.getColumns().size() == 3);
+		assertEquals("not", reader.getColumns().get(0));
+		assertEquals("a", reader.getColumns().get(1));
+		assertEquals("comment", reader.getColumns().get(2));
+
+		reader.readRow();
+		assertTrue(reader.getColumns().size() == 3);
+		assertEquals("also", reader.getColumns().get(0));
+		assertEquals("not", reader.getColumns().get(1));
+		assertEquals("comment", reader.getColumns().get(2));
+
+		assertFalse(reader.readRow());
+
+		reader.close();
+	}
+
+	/**
+	 * Tests that the readRow() method doesn't skip over empty lines if the ignoreEmptyLines
+	 * preference is disabled.
+	 */
+	@Test
+	public void testEmptyLinesWithIgnoreEmptyLines() throws Exception {
+
+		final CsvPreference notIgnoreEmptyLinesPrefs = new CsvPreference.Builder(CsvPreference.EXCEL_PREFERENCE)
+				.ignoreEmptyLines(false).build();
+
+		final String input = "\nthis is the second line\n\n";
+		AbstractCsvReader reader = new MockCsvReader(new StringReader(input), notIgnoreEmptyLinesPrefs);
+
+		reader.readRow();
+		assertTrue(reader.getColumns().size() == 1);
+		assertNull(reader.getColumns().get(0));
+		assertEquals(1, reader.getLineNumber());
+		assertEquals("", reader.getUndecodedRow());
+
+		reader.readRow();
+		assertTrue(reader.getColumns().size() == 1);
+		assertEquals("this is the second line", reader.getColumns().get(0));
+		assertEquals(2, reader.getLineNumber());
+		assertEquals("this is the second line", reader.getUndecodedRow());
+
+		reader.readRow();
+		assertTrue(reader.getColumns().size() == 1);
+		assertNull(reader.getColumns().get(0));
+		assertEquals(3, reader.getLineNumber());
+		assertEquals("", reader.getUndecodedRow());
+
+		reader.close();
+	}
+
+	/**
+	 * Tests the readRow() method when EOF is reached within quote scope.
+	 */
+	@Test
+	public void testQuotedFieldWithUnexpectedEOF() throws Exception {
+
+		// EOF reached within quote scope
+		final String input = "\"quoted spanning\ntwo lines with EOF reached before another quote";
+		AbstractCsvReader reader = new MockCsvReader(new StringReader(input), CsvPreference.EXCEL_PREFERENCE);
+		try {
+			reader.readRow();
+			fail("should have thrown SuperCsvException");
+		}
+		catch(SuperCsvException e) {
+			assertEquals("unexpected end of file while reading quoted column beginning on line 1 and ending on line 2",
+					e.getMessage());
+		}
+		reader.close();
+	}
+
+	/**
+     * Tests that the readRow() method skips over empty lines.
+     */
+    @Test
+    public void testEmptyLines() throws Exception {
+        final String input = "\n\nthis is the third line\n";
+        AbstractCsvReader reader = new MockCsvReader(new StringReader(input), CsvPreference.EXCEL_PREFERENCE);
+        reader.readRow();
+        assertTrue(reader.getColumns().size() == 1);
+        assertEquals("this is the third line", reader.getColumns().get(0));
+        assertEquals(3, reader.getLineNumber());
+        assertEquals("this is the third line", reader.getUndecodedRow());
+        reader.close();
+    }
+
+    /**
+     * Tests the readRow() method with an odd number of escape characters
+     * char.
+     */
+    @Test(expected = SuperCsvException.class)
+    public void testOddSeriesOfEscapeChars() throws Exception {
+
+        final CsvPreference csvPref = new CsvPreference.Builder(CsvPreference.STANDARD_PREFERENCE)
+                .setQuoteEscapeChar('#')
+                .build();
+
+        final String input = "\"#####\"";
+        abstractReader = new MockCsvReader(new StringReader(input), csvPref);
+        abstractReader.readRow();
+    }
+
+    /**
+     * Test double-quote char when in backslash-escape mode should throw exception
+     */
+    @Test(expected = SuperCsvException.class)
+    public void testDoubleQuoteBackslashEscapeChar() throws Exception {
+
+        // quote char is ' and escape char is $
+        final CsvPreference csvPref = new CsvPreference.Builder('\'', ',', "\n")
+                .setQuoteEscapeChar('$')
+                .build();
+
+        final String input = "'field with an escaped quote #' and a '' double quote'";
+        abstractReader = new MockCsvReader(new StringReader(input), csvPref);
+        abstractReader.readRow();
+
+        Assert.fail();
+    }
 }
